Dang particle spawner
This is not working as intended... There is a good chance that the
spawner is being rendered way off screen. Debugging information
indicates that it IS spawning particles, but they are nowhere to be
seen. Also, it appears the sprite is correct for the particles, so
location seems to be most likely the issue.